Fix funky taxonomy listing behaviour.
Basically, I was using get_terms() while attempting to get the taxonomy terms associated with my CPT.
At first, with minimal data, this seemed to work, except then I realised <duh> that just gets all the terms
that exist ever and that's going to be totally useless and super confusing. So if we use wp_get_post_terms()
instead and pass the post ID, everything is hunky-dory and it only returns the terms attached to our post.

Hurray!