Configure timezone input as an autocomplete.
In order to ensure we have consistent data for a person's timezone, we're using a list of PHP-supported timezones.
This should make it easier for PHP to handle the output of the field.

The timezone list as output by PHP is suuuuuper long, so I've configured this as an autocomplete for the moment.
This works nicely if you're in a major city, but is obviously less ideal if you're not. (For instance, if I enter
"London" it gives me London, but "Edinburgh" won't work.) Guessing in these cases is often easier with a giant
select box, but an autocomplete is easier to use if you know the best major city nearby to code it to.

Additionally, the output is currently using underscores, so a search for "New York" won't return any results.
Pretty easy fix if we decide an autocomplete is the better option here, although I'm not convinced it is.

Fixes #21.