Modify filter dropdown markup to accomodate hierarchical structures.
This turned out to be roughly 300x more complex than I was expecting, so I'm
basically having to make a bunch of arrays and loop through them. Right now,
this outputs all submenus as individual menus, outside of the main menu.

We'll be using this markup to animate the submenus in and out, assuming
I ever sort out how to make that work.