If a filter link has a sub-menu, make entire link click target.
Before, you could select the whole parent taxonomy by clicking it, and also
click through to the sub-menu by clicking the forward caret. This feels a bit
counter-intuitive though, so I've made the entire link open the sub-menu.

It means if you want to show all projects that are self-hosted, you would need
to click through each sub-menu item individually to turn them on, which could
be annoying. We can build a "show all" into the submenu as the first option
to make this easier, if it turns out we need that functionality.